+++ src/test/java/ladder/view/output/LadderCreateOutputViewTest.java
@@ -1,15 +1,15 @@
-package ladder.view;
+package ladder.view.output;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
 import static org.assertj.core.api.Assertions.*;
 
-class OutputViewTest {
+class LadderCreateOutputViewTest {
 
     @ParameterizedTest(name = "사람 이름이 5자 미만일 경우 공백을 채워 5자로 출력한다.")
     @ValueSource(strings = {"a","ab","abc","abcd","abcde"})
     void personname_output(String text) {
-        assertThat(OutputView.outputPersonName(text).length()).isEqualTo(5);
+        assertThat(LadderCreateOutputView.outputPersonName(text).length()).isEqualTo(5);
     }
 }
